Initial commit
This code is for the ESP8266 wifi module, which sends your data to IoT platforms. We have used ThingSpeak channel for the same. One important issue with ESP8266 is that you can send values of only one analog sensor, so we implemented multiplexing via HC 4051 IC to send 4 analog values one by one.